requirements.txt
spotipy==2.22.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
umap-learn==0.5.3
plotly==5.15.0
streamlit==1.25.0
python-dotenv==1.0.0
